Fixed a bug where the api returned doubles when the model expected ints.

+++ Plugin-Weather/WeatherLookupClasses.cs
@@ -8,14 +8,14 @@ namespace Plugin_Weather
         public int time { get; set; }
         public string summary { get; set; }
         public string icon { get; set; }
-        public int precipIntensity { get; set; }
-        public int precipProbability { get; set; }
+        public double precipIntensity { get; set; }
+        public double precipProbability { get; set; }
         public double temperature { get; set; }
         public double apparentTemperature { get; set; }
         public double dewPoint { get; set; }
         public double humidity { get; set; }
         public double windSpeed { get; set; }
-        public int windBearing { get; set; }
+        public double windBearing { get; set; }
         public double cloudCover { get; set; }
         public double pressure { get; set; }
         public double ozone { get; set; }
@@ -50,7 +50,7 @@ namespace Plugin_Weather
         public double dewPoint { get; set; }
         public double humidity { get; set; }
         public double windSpeed { get; set; }
-        public int windBearing { get; set; }
+        public double windBearing { get; set; }
         public double cloudCover { get; set; }
         public double pressure { get; set; }
         public double ozone { get; set; }
